% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_image.R
\name{geom_image}
\alias{geom_image}
\title{Make an image as ggplot2 background or change a bitmap to a ggplot2 object}
\usage{
geom_image(image)
}
\arguments{
\item{image}{raster object to display}
}
\value{

}
\description{
Make an image as ggplot2 background or change a bitmap to a ggplot2 object
}
\examples{
image <- jpeg::readJPEG(source = "Einstein.JPG")
p <- ggplot() + geom_image(image = image)
p <- p + ggpatch:::theme_nothing()
library(patchwork)
p1 <- ggplot(mtcars) + geom_point(aes(mpg, disp))
p + patchwork::inset_element(p1, 0.6, 0.6, 1, 1)
p1 + inset_element(p, 0.6, 0.6, 1, 1)
p + p1
image2 <- jpeg::readJPEG(source = "GoldenGirls.JPG")
p2 <- ggplot() + geom_image(image = image2) + ggpatch:::theme_nothing()
p + p1/p2 + patch_annotation()
}
\author{
Zhonghui Gai
}
