% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_tag}
\alias{patch_tag}
\title{Format the tag test of Figures, using with patchwork usually}
\usage{
patch_tag(vjust = -1)
}
\arguments{
\item{vjust}{adjust the y_position of the tags}
}
\value{

}
\description{
Format the tag test of Figures, using with patchwork usually
}
\examples{
library(ggplot2)
ggplot(mtcars) + geom_point(aes(mpg, disp)) + patch_tag()
}
\author{
Zhonghui Gai
}
